moved __tests__ folder to src folder as it contains tests for all FE components
