+++ src/__tests__/NavBarTest.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { render, screen } from '@testing-library/react';
 import { BrowserRouter } from "react-router-dom";
-import NavBar from '../NavBar';
+import NavBar from '../components/NavBar';
 import '@testing-library/jest-dom';
 
 describe('NavBar Component', () => {
